Refactor EncryptKey
